Refactored Input and Container files
