+++ jest.config.js
@@ -4,4 +4,5 @@ module.exports = {
     '\\.css$': require.resolve('./test/test-mock.js'),
     '\\.jpeg$': require.resolve('./test/test-mock.js'),
   },
+  setupTestFrameworkScriptFile: require.resolve('./test/setup-tests.js'),
 };
