+++ src/__tests__/componentRenders.js
@@ -1,4 +1,3 @@
-import 'react-testing-library/cleanup-after-each';
 import React from 'react';
 import { render } from 'react-testing-library';
