+++ test/setup-tests.js
@@ -1,0 +1,1 @@
+import 'react-testing-library/cleanup-after-each';
